requirements.txt
Flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
python-multipart==0.0.6
gunicorn==21.2.0 